+++ hwloc/topology-synthetic.c
@@ -60,7 +60,7 @@ hwloc_synthetic_process_level_indexes(struct hwloc_synthetic_backend_data_s *dat
   unsigned long length = curlevel->index_string_length;
   unsigned *array = NULL;
   struct hwloc_synthetic_intlv_loop_s * loops = NULL;
-  unsigned long i;
+  size_t i;
 
   if (!attr)
     return;
@@ -193,7 +193,7 @@ hwloc_synthetic_process_level_indexes(struct hwloc_synthetic_backend_data_s *dat
 	      && attrs.group.depth != (unsigned) -1
 	      && attrs.group.depth != data->level[i].depth)
 	    continue;
-	  loops[cur_loop].level_depth = i;
+	  loops[cur_loop].level_depth = (unsigned)i;
 	  break;
 	}
 	if (i == curleveldepth) {
@@ -321,7 +321,7 @@ hwloc_synthetic_parse_level_attrs(const char *attrs, const char **next_posp,
   const char *next_pos;
   hwloc_uint64_t memorysize = 0;
   const char *index_string = NULL;
-  unsigned long index_string_length = 0;
+  size_t index_string_length = 0;
 
   next_pos = (const char *) strchr(attrs, ')');
   if (!next_pos) {
@@ -365,7 +365,7 @@ hwloc_synthetic_parse_level_attrs(const char *attrs, const char **next_posp,
 
   curlevel->memorysize = memorysize;
   curlevel->index_string = index_string;
-  curlevel->index_string_length = index_string_length;
+  curlevel->index_string_length = (unsigned long)index_string_length;
   *next_posp = next_pos+1;
   return 0;
 }
@@ -1015,7 +1015,7 @@ static int hwloc_topology_export_synthetic_indexes(struct hwloc_topology * topol
     }
     ret += res;
     if (res >= tmplen)
-      res = tmplen>0 ? tmplen - 1 : 0;
+      res = tmplen>0 ? (int)tmplen - 1 : 0;
     tmp += res;
     tmplen -= res;
   }
@@ -1038,7 +1038,7 @@ static int hwloc_topology_export_synthetic_indexes(struct hwloc_topology * topol
       return -1;
     ret += res;
     if (res >= tmplen)
-      res = tmplen>0 ? tmplen - 1 : 0;
+      res = tmplen>0 ? (int)tmplen - 1 : 0;
     tmp += res;
     tmplen -= res;
     cur = cur->next_cousin;
@@ -1086,7 +1086,7 @@ static int hwloc_topology_export_synthetic_obj_attr(struct hwloc_topology * topo
       return -1;
     ret += res;
     if (res >= tmplen)
-      res = tmplen>0 ? tmplen - 1 : 0;
+      res = tmplen>0 ? (int)tmplen - 1 : 0;
     tmp += res;
     tmplen -= res;
 
@@ -1096,7 +1096,7 @@ static int hwloc_topology_export_synthetic_obj_attr(struct hwloc_topology * topo
 	return -1;
       ret += res;
       if (res >= tmplen)
-	res = tmplen>0 ? tmplen - 1 : 0;
+	res = tmplen>0 ? (int)tmplen - 1 : 0;
       tmp += res;
       tmplen -= res;
 
@@ -1105,7 +1105,7 @@ static int hwloc_topology_export_synthetic_obj_attr(struct hwloc_topology * topo
 	return -1;
       ret += res;
       if (res >= tmplen)
-	res = tmplen>0 ? tmplen - 1 : 0;
+	res = tmplen>0 ? (int)tmplen - 1 : 0;
       tmp += res;
       tmplen -= res;
     }
@@ -1160,7 +1160,7 @@ hwloc_topology_export_synthetic(struct hwloc_topology * topology,
     if (ret > 0)
       prefix = separator;
     if (res >= tmplen)
-      res = tmplen>0 ? tmplen - 1 : 0;
+      res = tmplen>0 ? (int)tmplen - 1 : 0;
     tmp += res;
     tmplen -= res;
   }
@@ -1180,7 +1180,7 @@ hwloc_topology_export_synthetic(struct hwloc_topology * topology,
       return -1;
     ret += res;
     if (res >= tmplen)
-      res = tmplen>0 ? tmplen - 1 : 0;
+      res = tmplen>0 ? (int)tmplen - 1 : 0;
     tmp += res;
     tmplen -= res;
 
@@ -1191,7 +1191,7 @@ hwloc_topology_export_synthetic(struct hwloc_topology * topology,
 	return -1;
       ret += res;
       if (res >= tmplen)
-	res = tmplen>0 ? tmplen - 1 : 0;
+	res = tmplen>0 ? (int)tmplen - 1 : 0;
       tmp += res;
       tmplen -= res;
     }
